requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
prisma==0.11.0
Werkzeug==3.0.1
nest-asyncio==1.5.8